requirements.txt
# Forge — Engineering Intelligence System
# Framework dependencies

# LLM + Orchestration
groq>=0.9.0
langgraph>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0

# Observability
wandb>=0.17.0

# Code Validation
ruff>=0.4.0
pylint>=3.2.0
bandit>=1.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-json-report>=1.5.0

# Dashboard UI
gradio>=4.40.0

# Utilities
pyyaml>=6.0
httpx>=0.27.0
tenacity>=8.3.0
python-jose>=3.3.0
passlib>=1.7.4

# Typing
typing-extensions>=4.12.0
